Paracou6_2016

Tree inventory of plot 6 of Paracou research station (French Guiana) for 2016 Public dataset extracted from the Guyafor database.

Paracou6_2016

Format

A tibble with 3620 rows and 26 variables:

Forest

Forest name (character)

Plot

Plot number (character)

PlotArea

Plot area, in hectare (numeric)

SubPlot

Subplot number (integer)

idTree

Unique tree identifier from the database (integer)

Protocole

Protocole name (character)

Xfield

Euclidean position of the tree relative to the Southwestern corner of the subplot, in m (numeric)

Yfield

Euclidean position of the tree relative to the Southwestern corner of the subplo, in m (numeric)

Xutm

Coordinate X for the tree in UTM 22 N (EPSG: 32 622) (numeric)

Yutm

Coordinate Y for the tree in UTM 22 N (EPSG: 32 622) (numeric)

UTMZone

UTM Zone (integer)

Lat

Tree latitude (WGS 84, EPSG : 4326) (numeric)

Lon

Tree longitude (WGS 84, EPSG : 4326) (numeric)

Family

Botanical family (character)

Genus

Botanical genus (character)

Species

Botanical species (character)

BotaSource

Source of botanical name (character)

BotaCertainty

Level of certainty for the botanical identification (numeric)

VernName

Vernacular name (character)

CensusYear

Census year (integer)

CensusDate

Census Day (when available) (character)

CensusDateCertainty

Is the CensusDate precise (TRUE) or not (FALSE)? If FALSE, this information must not be considered for calculation (logical)

CodeAlive

Is the tree alive (TRUE) or dead (FALSE)? (logical)

MeasCode

Information on the method for measuring the circumference or on the state of the tree (integer)

Circ

Circumference of the tree at 1.30m above ground (DBH level), in cm (numeric)

CircCorr

Corrected circumference (mathematical correction of abnormal circumferences), in cm (numeric)

CorrCode

Information on the reason why and how the circumference has been corrected, multiple values are possible(character)

...

Source

http://paracou.cirad.fr ; https://paracoudata.cirad.fr/public/pdf/Paracou_data_dictionnary.pdf

EcoFoG::Guyafor2df( WHERE = "Forest='Paracou' AND Plot='6' AND CensusYear=2016", UID = NULL,PWD = NULL,Driver = "SQL Server Native Client 10.0") %>% dplyr::select(-TreeFieldNum,-Project, -idVern, -CommercialSp, -CensusDate)