ParamCrownDiameterAllometry

Parameters of crown diameter allometry used to compute the crown diameter of each tree, depending on its DBH (Diameter at Breast Height) and its species, genus or family (Aubry-Kientz et al.2019).

ParamCrownDiameterAllometry

Format

A tibble with 371 rows and 7 variables:

alpha

Coefficient of the equation: ln(DBH) = alpha + beta ln(H*CD) + error, with error~N(0,sigma^2). DBH: tree diameter at breast hight, H: tree height, CD: tree crown diameter. Mean sigma^2 = 0.0295966977 (double)

beta

Coefficient of the equation: ln(DBH) = alpha + beta ln(H*CD) + error, with error~N(0,sigma^2).DBH: tree diameter at breast hight, H: tree height, CD: tree crown diameter. Mean sigma^2 = 0.0295966977 (double)

Taxo

Taxonomic level of the parameters (character)

ScientificName

Scientific name (character)

Genus

Genus name (character)

Species

Species name (character)

Family

Family name (character)

...