Directional felling success definition

Defines whether the directed fall of the tree is successful or not by drawing in a Bernoulli distribution where the probability of success is by default 60%, and can be changed with the advancedloggingparameters argument.

directionalfellingsuccessdef(
  inventory,
  fuel,
  directionalfelling,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory
input inventory (see the inputs formats and metadata in the vignette) (data.frame)

fuel
Fuel wood exploitation: no exploitation = "0", exploitation of damage and unused part of logged trees for fuel = "1", exploitation of hollow trees, damage and and unused part of the log for fuel = "2"

directionalfelling
Directional felling = "0": only to direct the foot of the tree towards the trail "1": to direct the foot of the tree towards the trail + to avoid damage to future and reserve trees "2": to avoid damage to future and reserve trees + orientation angle to the trail

advancedloggingparameters
Other parameters of the logging simulator loggingparameters (list)

Value

Input inventory with the "TreeFellingOrientationSuccess" new column (see the outputs metadata in the vignette).

Examples

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- treeselection(inventory, objective = 20, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = TRUE, topography = DTMParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory

#> As the harvestable volume (= 200.6m^3)
#>                 was higher (by 75.6m^3)
#>                 than the objective volume, it was necessary to increase the Minimum Falling Diameter (MinFD)
#>                 of all species. The objective volume has now been reached.

new <- directionalfellingsuccessdef(inventory,fuel = "2", directionalfelling = "2",
advancedloggingparameters = loggingparameters())