Future & reserve trees designation

Future & reserve trees designation

futurereserve(
  inventory,
  speciescriteria,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory
Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

speciescriteria
Table of species exploitability criteria : species names, economic interest level, minimum and maximum felling diameter, in the same format as SpeciesCriteria (data.frame)

advancedloggingparameters
Other parameters of the logging simulator loggingparameters (list)

Value

input inventory with selected future and reserve trees (LoggingStatus = "future"/"reserve"), for your logging criteria (SpeciesCriteria)

Details

Future trees are all trees satisfying the following conditions:

species of 1st economic rank

DBH between 35cm ('FutureTreesMinDiameter') and the species MinFD or UpMinFD if it has been raised for its species.

Reserve trees are randomly chosen among future trees so that the number of reserve trees is equal to the number of harvested trees.

See also

Examples

data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(SpeciesCriteria)
data(DTMParacou)
data(PlotSlope)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
topography = DTMParacou, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, scenario = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = TRUE,
topography = DTMParacou,
advancedloggingparameters = loggingparameters(), VO = 125,
 HVinit = HVinit)$inventory

#> As the harvestable volume (= 199.7m^3)
#>                 was higher (by 74.7m^3)
#>                 than the objective volume, it was necessary to increase the Minimum Falling Diameter (MinFD)
#>                 of all species. The objective volume has now been reached.

futurereserve(inventory, SpeciesCriteria)

#> # A tibble: 3,539 x 54
#>    Forest  Plot  PlotArea SubPlot idTree Protocole Xfield Yfield    Xutm    Yutm
#>    <chr>   <chr>    <dbl>   <int>  <int> <chr>      <dbl>  <dbl>   <dbl>   <dbl>
#>  1 Paracou 6         6.25       1 101474 Guyafor    106.   238.  286500. 583194.
#>  2 Paracou 6         6.25       1 100727 Guyafor     12.5  200.  286419. 583135.
#>  3 Paracou 6         6.25       2 102482 Guyafor    232    246   286618. 583231.
#>  4 Paracou 6         6.25       3 103625 Guyafor    111     68.5 286546. 583029.
#>  5 Paracou 6         6.25       3 102824 Guyafor      5.5   24.5 286456. 582961.
#>  6 Paracou 6         6.25       4 104246 Guyafor    176.     2.5 286625. 582980.
#>  7 Paracou 6         6.25       2 102520 Guyafor    241    181   286642. 583170.
#>  8 Paracou 6         6.25       2 101862 Guyafor    142.   173   286550. 583138.
#>  9 Paracou 6         6.25       3 103693 Guyafor    125    109   286549. 583072.
#> 10 Paracou 6         6.25       3 103544 Guyafor    105     20   286552. 582980.
#> # ... with 3,529 more rows, and 44 more variables: UTMZone <int>, Lat <dbl>,
#> #   Lon <dbl>, Family <chr>, ScientificName <chr>, Genus <chr>, Species <chr>,
#> #   BotaSource <chr>, BotaCertainty <dbl>, VernName <chr>, CensusYear <int>,
#> #   CensusDateCertainty <lgl>, CodeAlive <lgl>, MeasCode <int>, Circ <dbl>,
#> #   CircCorr <dbl>, CorrCode <chr>, DBH <dbl>, TreeHarvestableVolume <dbl>,
#> #   TrunkHeight <dbl>, TreeHeight <dbl>, CrownHeight <dbl>, Taxo <chr>,
#> #   CrownDiameter <dbl>, WoodDensity <dbl>, levelWD <chr>, AGB <dbl>,
#> #   ONFName <chr>, Commercial <fct>, MinFD <dbl>, UpMinFD <dbl>, MaxFD <dbl>,
#> #   Aggregative <lgl>, DistCrit <lgl>, Slope <dbl>, SlopeCrit <lgl>,
#> #   LoggingStatus <chr>, LogDBH <dbl>, VisibleDefectProba <dbl>,
#> #   VisibleDefect <fct>, ProbedHollowProba <dbl>, ProbedHollow <fct>,
#> #   Selected <chr>, Up <chr>