Advanced parameters of the logging simulator

Advanced parameters of the logging simulator

loggingparameters(
  MinDBHValue = 10,
  MaxAreaSlope = 27,
  MaxTrailCenterlineSlope = 22,
  MaxTrailCrossSlope = 4,
  GrappleMaxslope = 20,
  TreeMaxSlope = 22,
  PlateauMaxSlope = 5,
  SlopeDistance = 6,
  WaterSourcesBufferZone = 30,
  MinMainTrailWidth = 5,
  MaxMainTrailWidth = 6,
  ScndTrailWidth = 4,
  BigTrees = 50,
  CableLength = 40,
  GrappleLength = 6,
  IsolateTreeMinDistance = 100,
  FutureTreesMinDiameter = 35,
  TreefallSuccessProportion = 0.6,
  MinTreefallOrientation = 30,
  MaxTreefallOrientation = 45,
  TreeHollowPartForFuel = 1/3,
  Purge = 0.14,
  MaxTrailDensity = 200,
  MaxLandingArea = 1500,
  TreeHarvestableVolumeAllometry = function(DBH, aCoef, bCoef) aCoef + bCoef *
    (DBH/100)^2,
  TrunkHeightAllometry = function(DBH, TreeHarvestableVolume) TreeHarvestableVolume/(pi
    * (((DBH/100)/2)^2)),
  TreeHeightAllometry = function(DBH) exp(0.07359191 + 1.34241216 * log(DBH) +
    -0.12282344 * log(DBH)^2),
  CrownDiameterAllometry = function(DBH, TreeHeight, alpha, beta) exp(((log(DBH) -
    alpha - rnorm(length(DBH), 0, 0.0295966977))/beta))/TreeHeight,
  RottenModel = function(DBH) 1/(1 + exp(-(-5.151 + 0.042 * DBH))),
  VisiblyDefectModel = function(LogDBH) 1/(1 + exp(-(-3.392 + 0.357 * LogDBH)))
)

Arguments

MinDBHValue
Minimum DBH for inclusion in the forest inventory. Default = 10, in cm (double)

MaxAreaSlope
Maximum area slope for inclusion in a logging unit. Default = 27, in % (double)

MaxTrailCenterlineSlope
Maximum main & 2nd trail centerline slope. Default = 22, in % (double)

MaxTrailCrossSlope
Maximum main & 2nd trail cross slope. Default = 4, in % (double)

GrappleMaxslope
Maximum slope accessible by the grapple. Default = 20, in % (double)

TreeMaxSlope
Maximum slope around the tree to access it. Default = 22, in % (double)

PlateauMaxSlope
Maximum slope to define an area as a plateau. Default = 5, in % (double)

SlopeDistance
Distance over which the slope is calculated. Default = 6, in m (3m each side) (double)

WaterSourcesBufferZone
Buffer zone around the water sources. Default = 30, in m (double)

MinMainTrailWidth
Minimum main trail width. Default = 5, in m (double)

MaxMainTrailWidth
Maximum main trail width. Default = 6, in m (double)

ScndTrailWidth
2nd trail width. Default = 4, in m (double)

BigTrees
Minimum DBH of trees to be avoided by trails. Default = 50, in cm (double)

CableLength
Cable length. Default = 40, in m (double)

GrappleLength
Grapple length. Default = 6, in m (double)

IsolateTreeMinDistance
Minimum distance to consider a tree "isolated" from other trees of its species, in the aggregative species case (SpeciesCriteria, 'Aggregative' column). Default = 100, in m (double)

FutureTreesMinDiameter
Future trees minimum diameter. Default = 35, in cm (double)

TreefallSuccessProportion
Proportion of successful directional felling events. Default = 0.6 (double)

MinTreefallOrientation
Minimum orientation of the tree fall to the trail. Default = 30, in degree (double)

MaxTreefallOrientation
Maximum orientation of the tree fall to the trail. Default = 45, in degree (double)

TreeHollowPartForFuel
Proportion of hollow trees used as fuel wood . Default = 1/3 (double)

Purge
Part of the log no used for timber, can be used for fuel wood. Default = 0.14, in m^3 of purge/m^3 of volume of timber harvested. (double)

MaxTrailDensity
Maximum trail density. Default = 200, in m/ha (double) (has no impact on the simulation. A message will be sent to inform if this threshold has been exceeded)

MaxLandingArea
Maximum landing area. Default = 1500) in square meters (double) (has no impact on the simulation. A message will be sent to inform if this threshold has been exceeded)

TreeHarvestableVolumeAllometry
By default, allometry of tree harvestable volume, French Guiana ONF formula: aCoef + bCoef * (DBH/100)^2, aCoef and bCoef depend on the forest location, stored in ForestZoneVolumeParametersTable, DBH in cm. (function)

TrunkHeightAllometry
Allometry of trunk height, based on the cylinder volume formula: CylinderVolume = pi ((DBH/100)/2)^2 * H, with the height (H) in m and the DBH in cm (function)

TreeHeightAllometry
By default, allometry parameters estimated from Guyanese data with the BIOMASS package: ln(H) = 0.07359191 + 1.34241216 * ln(DBH) -0.12282344 * ln(DBH)^2, with the height (H) in m and the DBH in cm (function)

CrownDiameterAllometry
ln(DBH) = 𝜶+ 𝜷 ln(H*CD) + 𝜺, with 𝜺~N(0,σ^2) and mean σ^2 = 0.0295966977 with the crown diameter (CD), the tree height (H) in m, and the DBH in cm. (Aubry-Kientz et al.2019)(function)

RottenModel
Estimates the tree probability of being probed hollow (default: 1 / (1 + exp(-(-5.151 + 0.042 DBH))) with DBH in cm) (function)

VisiblyDefectModel
Estimates the tree probability to have visible defects. Default: 1 / (1 + exp(-(-3.392 + 0.357 * ln(DBH)))) with DBH in cm (function)

Value

A named list of 30 objects.

Examples

loggingparameters(MinDBHValue = 5)

#> $MinDBHValue
#> [1] 5
#> 
#> $MaxAreaSlope
#> [1] 27
#> 
#> $MaxTrailCenterlineSlope
#> [1] 22
#> 
#> $MaxTrailCrossSlope
#> [1] 4
#> 
#> $GrappleMaxslope
#> [1] 20
#> 
#> $TreeMaxSlope
#> [1] 22
#> 
#> $PlateauMaxSlope
#> [1] 5
#> 
#> $SlopeDistance
#> [1] 6
#> 
#> $WaterSourcesBufferZone
#> [1] 30
#> 
#> $MinMainTrailWidth
#> [1] 5
#> 
#> $MaxMainTrailWidth
#> [1] 6
#> 
#> $ScndTrailWidth
#> [1] 4
#> 
#> $BigTrees
#> [1] 50
#> 
#> $CableLength
#> [1] 40
#> 
#> $GrappleLength
#> [1] 6
#> 
#> $IsolateTreeMinDistance
#> [1] 100
#> 
#> $FutureTreesMinDiameter
#> [1] 35
#> 
#> $TreefallSuccessProportion
#> [1] 0.6
#> 
#> $MinTreefallOrientation
#> [1] 30
#> 
#> $MaxTreefallOrientation
#> [1] 45
#> 
#> $TreeHollowPartForFuel
#> [1] 0.3333333
#> 
#> $Purge
#> [1] 0.14
#> 
#> $MaxTrailDensity
#> [1] 200
#> 
#> $MaxLandingArea
#> [1] 1500
#> 
#> $TreeHarvestableVolumeAllometry
#> function (DBH, aCoef, bCoef) 
#> aCoef + bCoef * (DBH/100)^2
#> <environment: 0x000000008610c440>
#> 
#> $TrunkHeightAllometry
#> function (DBH, TreeHarvestableVolume) 
#> TreeHarvestableVolume/(pi * (((DBH/100)/2)^2))
#> <environment: 0x000000008610c440>
#> 
#> $TreeHeightAllometry
#> function (DBH) 
#> exp(0.07359191 + 1.34241216 * log(DBH) + -0.12282344 * log(DBH)^2)
#> <environment: 0x000000008610c440>
#> 
#> $CrownDiameterAllometry
#> function (DBH, TreeHeight, alpha, beta) 
#> exp(((log(DBH) - alpha - rnorm(length(DBH), 0, 0.0295966977))/beta))/TreeHeight
#> <environment: 0x000000008610c440>
#> 
#> $RottenModel
#> function (DBH) 
#> 1/(1 + exp(-(-5.151 + 0.042 * DBH)))
#> <environment: 0x000000008610c440>
#> 
#> $VisiblyDefectModel
#> function (LogDBH) 
#> 1/(1 + exp(-(-3.392 + 0.357 * LogDBH)))
#> <environment: 0x000000008610c440>
#>