Future & reserve trees designation

Future & reserve trees designation

Usage

futurereserve(
  inventory,
  speciescriteria,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

speciescriteria

Table of species exploitability criteria : species names, economic interest level, minimum and maximum felling diameter, in the same format as SpeciesCriteria (data.frame)

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

input inventory with selected future and reserve trees (LoggingStatus = "future"/"reserve"), for your logging criteria (SpeciesCriteria)

Details

Future trees are all trees satisfying the following conditions:

species of 1st economic rank

DBH between 35cm ('FutureTreesMinDiameter') and the species MinFD or UpMinFD if it has been raised for its species.

in the prospection units (harvestable areas)

Reserve trees are randomly chosen among future trees so that the number of reserve trees is equal to the number of harvested trees.

See also

Examples

data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(SpeciesCriteria)
data(DTMParacou)
data(HarvestableAreaOutputsCable)

scenario <- "RIL2"

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory,
 topography = DTMParacou,
 diversification = TRUE,
 specieslax = FALSE,
 plotslope = HarvestableAreaOutputsCable$PlotSlope,
 harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
 scenario = scenario,
 advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, scenario = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = TRUE,
topography = DTMParacou,
advancedloggingparameters = loggingparameters(), VO = 125,
 HVinit = HVinit)$inventory
#> As the harvestable volume (= 214m^3)
#>                 was higher (by 89m^3)
#>                 than the objective volume, the Minimum Falling Diameter (MinFD) of the 1st and 2nd economic ranks species
#>                   were increased to UpMinFD. The objective volume has now been reached.

futurereserve(inventory, SpeciesCriteria)
#> # A tibble: 3,525 × 56
#>    Forest  Plot  PlotArea SubPlot idTree Protocole Xfield Yfield UTMZone   Lat
#>    <chr>   <chr>    <dbl>   <int>  <int> <chr>      <dbl>  <dbl>   <int> <dbl>
#>  1 Paracou 6         6.25       2 102249 Guyafor    202    134.       22  5.27
#>  2 Paracou 6         6.25       2 102086 Guyafor    176.   182.       22  5.27
#>  3 Paracou 6         6.25       1 100727 Guyafor     12.5  200.       22  5.27
#>  4 Paracou 6         6.25       4 104642 Guyafor    242     31.5      22  5.27
#>  5 Paracou 6         6.25       1 101436 Guyafor    108.   170        22  5.27
#>  6 Paracou 6         6.25       1 101413 Guyafor    106.   140.       22  5.27
#>  7 Paracou 6         6.25       3 102824 Guyafor      5.5   24.5      22  5.27
#>  8 Paracou 6         6.25       1 100699 Guyafor     20    240.       22  5.27
#>  9 Paracou 6         6.25       1 101055 Guyafor     58    173        22  5.27
#> 10 Paracou 6         6.25       2 102507 Guyafor    234.   202.       22  5.27
#> # … with 3,515 more rows, and 46 more variables: Lon <dbl>, Family <chr>,
#> #   ScientificName <chr>, Genus <chr>, Species <chr>, BotaSource <chr>,
#> #   BotaCertainty <dbl>, VernName <chr>, CensusYear <int>,
#> #   CensusDateCertainty <lgl>, CodeAlive <lgl>, MeasCode <int>, Circ <dbl>,
#> #   CircCorr <dbl>, CorrCode <chr>, DBH <dbl>, TreesIn <chr>, Xutm <dbl>,
#> #   Yutm <dbl>, TreeHarvestableVolume <dbl>, TrunkHeight <dbl>,
#> #   TreeHeight <dbl>, CrownHeight <dbl>, Taxo <chr>, CrownDiameter <dbl>, …